Permissions of offer controller

+++ backend/src/main/java/com/uah/gestion_de_practicas/service/OfferService.java
@@ -2,8 +2,10 @@ package com.uah.gestion_de_practicas.service;
 
 import java.util.List;
 
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
+import com.uah.gestion_de_practicas.model.Company;
 import com.uah.gestion_de_practicas.model.Offer;
 import com.uah.gestion_de_practicas.repository.OfferRepository;
 
@@ -19,8 +21,20 @@ public class OfferService {
      */
     private final OfferRepository offerRepository;
 
-    public OfferService(OfferRepository offerRepository) {
+    /** 
+     * Service object for the Tutor class
+     */
+    private final TutorService tutorService;
+
+    /** 
+     * Service object for the Company class
+     */
+    private final CompanyService companyService;
+
+    public OfferService(OfferRepository offerRepository, TutorService tutorService, @Lazy CompanyService companyService) {
         this.offerRepository = offerRepository;
+        this.tutorService = tutorService;
+        this.companyService = companyService;
     }
 
     // ------------------- CRUD OPERATIONS ------------------- //
@@ -28,7 +42,10 @@ public class OfferService {
     /**
      * Saves an offer in the database.
      */
-    public Offer saveOffer(Offer offer) {
+    public Offer saveOffer(Offer offer, String tutor_username) {
+        if (!tutorService.isAuthorized(tutor_username, offer.getCompany().getId())) {
+            return null;
+        }
         return offerRepository.save(offer);
     }
 
@@ -43,9 +60,16 @@ public class OfferService {
     /**
      * Deletes an offer from the database.
      * @param id Id of the offer to be deleted.
+     * @param tutor_username Username of the tutor that wants to delete the offer.
+     * @return The boolean value true if the offer was deleted, false otherwise.
      */
-    public void deleteOffer(Long id) {
+    public boolean deleteOffer(Long id, String tutor_username) {
+        Offer offer = offerRepository.findById(id).orElse(null);
+        if (offer == null || !tutorService.isAuthorized(tutor_username, offer.getCompany().getId())) {
+            return false;
+        }
         offerRepository.deleteById(id);
+        return true;
     }
 
     /**
@@ -59,4 +83,12 @@ public class OfferService {
 
     // ------------------- OTHER METHODS ------------------- //
     
+    /** 
+     * Get the company of an offer
+     * @param id, the id of the company
+     * @return the company of the offer
+     */
+    public Company getCompany(Long id) {
+        return companyService.getCompany(id);
+    }
 }
